Second but first working commit.
This commit  contains a  huge amount  of rework  (and rethink)  on how
Chester should  work, what  it really  is all  about. I  had a  lot of
problem getting  the true server  idea to  work, mostly because  I was
hesitant towards  writing a  server from scratch;  it feels  like that
should not be necessary in this day of age.

In an act  of desperation I looked  at git's source code  in order to
figure out how  it can (or should) be done,  how to exchange encrypted
and authenticated data over our oh-so-beloved internet. It seemed like
ssh was to  way to go but it  took me a while to  realize how (reading
C-code is hard, so I had to mull it over for some time).

After  quite  some  time  I  succeeded  to  hijack  ssh  into  my  own
bidding.  The Chester  project was  reborn and  everything was  bliss,
truly. This commit contains a lot  of stuff, mostly bad stuff but with
the occasional  okay bits and pieces.  A lot of it  will most probably
change some time in the future.